homework_4 - done
*Вынес всю логику в класс TestFramework.java
*Создал в TestFramework.java метод runTest(), принимающий на вход класс теста и прогоняющий его
*Поменял логику работы фреймворка, теперь перед и после каждого метода аннотированого @Test прогоняются методы @Before и @After
