+++ lesson_8_ReflectionAnnotation/src/main/java/ru/sbrf/TestRunner.java
@@ -1,0 +1,9 @@
+package ru.sbrf;
+
+public class TestRunner {
+    public static void main(String[] args) {
+        TestFramework testFramework = new TestFramework();
+
+        testFramework.runTest(TestExample.class);
+    }
+}
